Fixes duplicate artworks in curator artwork page
pg_search creates a complicated long sql query when search is combined with other activerecord filters.
we want either `group by` artworks or have `distinct` artwork ids when checking for artwork images (in `with_images`) and the combination of `order by` for sorting by rank that pg_search adds and the distinct/group statement causes a SQL error because, in case of `distinct`, sql doesn't like the order by expression not be in the select expression. adding `with_pg_search_rank` (a feature in pg_search mostly for debugging) adds the rank to the select clause and makes sql happy