Fix lightbox problem
light box was opening the second image when clicked. I tried to debug lightbox code to see why but couldn't find anything. as a workaround added `pointer-events: none` to the images that are not displayed by carousel to make sure the click event only gets to the current lightbox container